Setup Wails as frontend

+++ setup.sh
@@ -26,7 +26,12 @@ main() {
     go_get github.com/wailsapp/wails/v2
 
     echo "=== Создание .env ==="
-    cp .env.example .env
+    if ![ -f ".env" ]; then
+        cp .env.example .env
+        echo ".env создан"
+    else
+        echo ".env уже существует"
+    fi
 
     echo "Настройка завершена."
 }
